feat(rbac): cover conditional API with tests. Improve code

+++ plugins/rbac-backend/src/service/policies-rest-api.ts
@@ -361,6 +361,9 @@ export class PolicesServer {
       }
 
       const id: number = parseInt(req.params.id, 10);
+      if (isNaN(id)) {
+        throw new InputError('Id is not a valid number.');
+      }
 
       const condition = await this.conditionalStorage.getCondition(id);
       if (!condition) {
@@ -385,6 +388,9 @@ export class PolicesServer {
       }
 
       const id: number = parseInt(req.params.id, 10);
+      if (isNaN(id)) {
+        throw new InputError('Id is not a valid number.');
+      }
 
       await this.conditionalStorage.deleteCondition(id);
       resp.status(204).end();
@@ -405,6 +411,9 @@ export class PolicesServer {
       }
 
       const id: number = parseInt(req.params.id, 10);
+      if (isNaN(id)) {
+        throw new InputError('Id is not a valid number.');
+      }
       const conditionalPolicy: ConditionalPolicyDecision = req.body;
 
       await this.conditionalStorage.updateCondition(id, conditionalPolicy);
